feat(sticker): Implementa Gerenciador de Pacotes de Figurinhas e Subcomandos
- Adicionado stickerPackManager.js para gerenciar pacotes de figurinhas dos usuários, incluindo funcionalidades para adicionar, listar, excluir, renomear pacotes e enviar figurinhas individualmente.
- Introduzido stickerSubCommands.js para lidar com subcomandos relacionados aos pacotes de figurinhas, como listar pacotes, exibir estatísticas e enviar pacotes completos.
- Melhorada a experiência do usuário com mensagens detalhadas e tratamento de erros nas operações de figurinhas.
- Implementado sistema de logs para melhor rastreabilidade de ações e erros.
